refactor(cli): extract rendering helpers to app/cli/render.py
Moves the 10 Rich-rendering helpers (_render_task_event,
_render_log_entry, _render_planned_table, _render_run_footer,
_truncate_tail, _truncated_section, _build_failure_body,
_task_status_summary, _FLOW_STATUS_STYLE, _TASK_STATUS_GLYPHS) out of
app/main.py into app/cli/render.py. app/main.py re-imports them so the
remaining command bodies still resolve, and so tests/test_main.py
continues to import _render_task_event/_truncate_tail from app.main
until Task 12.

Step 2 of the app/cli/ extraction. Tests stay green; --help byte-identical.
